refactor:응집도 개선(데이터 페칭 로직을 커스텀 훅으로 분리)
